fix code smell

+++ src/gui/PerformancePanel.java
@@ -18,7 +18,7 @@ import oshi.hardware.HWDiskStore;
 import oshi.hardware.NetworkIF;
 
 public class PerformancePanel extends OshiJPanel{
-  
+
     public PerformancePanel() {
         super();
     }
@@ -32,21 +32,21 @@ public class PerformancePanel extends OshiJPanel{
         }
         initial(si);
     }
-    
+
     class JVerticalMenuBar extends JMenuBar {
         private static final LayoutManager grid = new GridLayout(0,1);
         public JVerticalMenuBar() {
             setLayout(grid);
         }
     }
-    
+
     public static void setChartRenderer(JFreeChart chart, Color color){
         XYAreaRenderer renderer = new XYAreaRenderer();
         renderer.setSeriesPaint(0, new Color(color.getRed(), color.getGreen(), color.getBlue(), 128));
         renderer.setOutline(true);
         renderer.setSeriesOutlineStroke(0, new BasicStroke(1.0f));
         renderer.setSeriesOutlinePaint(0, color.darker());
-        
+
         chart.getXYPlot().setRenderer(renderer);
         chart.getPlot().setBackgroundPaint( Color.WHITE );
         chart.getXYPlot().setDomainGridlinesVisible(true);
@@ -61,7 +61,7 @@ public class PerformancePanel extends OshiJPanel{
         renderer.setOutline(true);
         renderer.setSeriesOutlineStroke(0, new BasicStroke(1.0f));
         renderer.setSeriesOutlinePaint(0, color1.darker());
-        
+
         renderer.setSeriesPaint(1, new Color(color2.getRed(), color2.getGreen(), color2.getBlue(), 128));
         renderer.setOutline(true);
         renderer.setSeriesOutlineStroke(1, new BasicStroke(1.0f));
@@ -90,7 +90,7 @@ public class PerformancePanel extends OshiJPanel{
 
         perfMenuBar.add(cpuButton);
         CPUPanel.updateCPUInfo(si.getHardware().getProcessor(), cpuButton, si.getOperatingSystem());
-        
+
         JGradientButton memButton = createButton(PerformancePanel.buttonTextLines("\nMemory\n0/0 GB (0%)\n"), 'M', "Display Memory", Color.GREEN,new MemoryPanel(si), displayPanel);
 
         perfMenuBar.add(memButton);
@@ -106,7 +106,7 @@ public class PerformancePanel extends OshiJPanel{
             perfMenuBar.add(diskButton[i]);
         }
         DiskPanel.updateDiskInfo(si.getHardware().getDiskStores(), diskButton);
-        
+
         List<NetworkIF> networkIFs = si.getHardware().getNetworkIFs(true);
 
         JGradientButton[] netButton = new JGradientButton[networkIFs.size()];
@@ -122,7 +122,7 @@ public class PerformancePanel extends OshiJPanel{
         scrollPerfPanel.getVerticalScrollBar().setUnitIncrement(30);
         scrollPerfPanel.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
         scrollPerfPanel.setHorizontalScrollBarPolicy(ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
-        
+
         scrollPerfPanel.setBounds(0, 0, 295, 935);
         perfPanel.add(scrollPerfPanel);
 
@@ -168,8 +168,8 @@ public class PerformancePanel extends OshiJPanel{
         button.setHorizontalAlignment(SwingConstants.LEFT);
         // Set what to do when we push the button
         button.addActionListener(e -> {
-            int nComponents = (int)displayPanel.getComponents().length;
-            if (nComponents <= (int)0 || displayPanel.getComponent(0) != panel) {
+            int nComponents = displayPanel.getComponents().length;
+            if (nComponents <= 0 || displayPanel.getComponent(0) != panel) {
                 resetMainGui(displayPanel);
                 displayPanel.add(panel);
                 refreshMainGui(displayPanel);
@@ -177,7 +177,7 @@ public class PerformancePanel extends OshiJPanel{
         });
         return button;
     }
-    
+
     public static void resetMainGui(JPanel displayPanel) {
         displayPanel.removeAll();
     }
